Final progress with inventor before scrapping for another solution
When I first started this I was going step by step but I think it's either impossible or too difficult or too ugly to interface a python script with inventor at the moment. Inventor's python API seems to be very lacking or at least there isn't much info on it. Instead I will be restarting the project, likely with something like FreeCAD. I don't love FreeCAD but that may be the best option.